+++ src/main/java/de/dm/intellij/bndtools/commenter/BndCommenter.java
@@ -1,0 +1,32 @@
+package de.dm.intellij.bndtools.commenter;
+
+import com.intellij.lang.Commenter;
+import org.jetbrains.annotations.Nullable;
+
+public class BndCommenter implements Commenter {
+
+    @Override
+    public @Nullable String getLineCommentPrefix() {
+        return "#";
+    }
+
+    @Override
+    public @Nullable String getBlockCommentPrefix() {
+        return null;
+    }
+
+    @Override
+    public @Nullable String getBlockCommentSuffix() {
+        return null;
+    }
+
+    @Override
+    public @Nullable String getCommentedBlockCommentPrefix() {
+        return null;
+    }
+
+    @Override
+    public @Nullable String getCommentedBlockCommentSuffix() {
+        return null;
+    }
+}